+++ treeshr/TreeGetDbi.c
@@ -14,7 +14,7 @@ int TreeGetDbi(struct dbi_itm *itmlst)
 
 #define set_retlen(length) if ((unsigned int)lst->buffer_length < length) { status = TreeBUFFEROVF; break; } else retlen=length
 #define CheckOpen(db) if (!db || !db->open) {status=TreeNOT_OPEN;break;}
-#define set_ret_char(val) memset(lst->pointer, 0, lst->buffer_length); *((char *)lst->pointer) = val
+#define set_ret_char(val) memset(lst->pointer, 0, (size_t)lst->buffer_length); *((char *)lst->pointer) = val
 
 int _TreeGetDbi(void *dbid, struct dbi_itm *itmlst)
 {
@@ -72,8 +72,8 @@ int _TreeGetDbi(void *dbid, struct dbi_itm *itmlst)
 	PINO_DATABASE *db_tmp;
 	for (count = 0, db_tmp = (PINO_DATABASE *) dbid; db_tmp ? db_tmp->open : 0;
 	     count++, db_tmp = db_tmp->next) ;
-	memset(lst->pointer, 0, lst->buffer_length);
-	memcpy(lst->pointer, &count, minInt(lst->buffer_length, sizeof(int)));
+	memset(lst->pointer, 0, (size_t)lst->buffer_length);
+	memcpy(lst->pointer, &count, (size_t)minInt(lst->buffer_length, sizeof(int)));
 	if (lst->return_length_address)
 	  *lst->return_length_address = minInt(lst->buffer_length, sizeof(int));
 	break;
@@ -83,8 +83,8 @@ int _TreeGetDbi(void *dbid, struct dbi_itm *itmlst)
 
       {
 	int count = db->stack_size;
-	memset(lst->pointer, 0, lst->buffer_length);
-	memcpy(lst->pointer, &count, minInt(lst->buffer_length, sizeof(int)));
+	memset(lst->pointer, 0, (size_t)lst->buffer_length);
+	memcpy(lst->pointer, &count, (size_t)minInt(lst->buffer_length, sizeof(int)));
 	if (lst->return_length_address)
 	  *lst->return_length_address = minInt(lst->buffer_length, sizeof(int));
 	break;
@@ -104,8 +104,8 @@ int _TreeGetDbi(void *dbid, struct dbi_itm *itmlst)
       CheckOpen(db);
       {
 	int value = db->tree_info->header->versions_in_model;
-	memset(lst->pointer, 0, lst->buffer_length);
-	memcpy(lst->pointer, &value, minInt(lst->buffer_length, sizeof(int)));
+	memset(lst->pointer, 0, (size_t)lst->buffer_length);
+	memcpy(lst->pointer, &value, (size_t)minInt(lst->buffer_length, sizeof(int)));
 	if (lst->return_length_address)
 	  *lst->return_length_address = minInt(lst->buffer_length, sizeof(int));
 	break;
@@ -115,8 +115,8 @@ int _TreeGetDbi(void *dbid, struct dbi_itm *itmlst)
       CheckOpen(db);
       {
 	int value = db->tree_info->header->versions_in_pulse;
-	memset(lst->pointer, 0, lst->buffer_length);
-	memcpy(lst->pointer, &value, minInt(lst->buffer_length, sizeof(int)));
+	memset(lst->pointer, 0, (size_t)lst->buffer_length);
+	memcpy(lst->pointer, &value, (size_t)minInt(lst->buffer_length, sizeof(int)));
 	if (lst->return_length_address)
 	  *lst->return_length_address = minInt(lst->buffer_length, sizeof(int));
 	break;
@@ -128,8 +128,8 @@ int _TreeGetDbi(void *dbid, struct dbi_itm *itmlst)
     }
     if (string) {
       if (lst->buffer_length && lst->pointer) {
-	retlen = minInt(strlen(string), lst->buffer_length);
-	strncpy((char *)lst->pointer, string, retlen);
+	retlen = (unsigned short)minInt((int)strlen(string), lst->buffer_length);
+	strncpy((char *)lst->pointer, string, (size_t)retlen);
 	if (retlen < lst->buffer_length)
 	  ((char *)lst->pointer)[retlen] = '\0';
 	free(string);
